requirements.txt
dotenv
azure-storage-blob
